backend/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==5.0.1
django-ninja==1.1.0
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
sqlparse==0.4.4
typing_extensions==4.9.0
